+++ src/main/fc/rc_modes.h
@@ -69,6 +69,7 @@ typedef enum {
     BOXUSER2,
     BOXUSER3,
     BOXUSER4,
+    BOXFFESTIMATE,
 
     CHECKBOX_ITEM_COUNT,