+++ src/main/pg/mixer.c
@@ -40,6 +40,11 @@ PG_RESET_TEMPLATE(mixerConfig_t, mixerConfig,
     .swash_geo_correction = 0,
     .collective_tilt_correction_pos = 0,
     .collective_tilt_correction_neg = 10,
+    .ff_estimate_gain = 50,
+    .ff_estimate_min_setpoint = 50,
+    .ff_estimate_convergence = 10,
+    .ff_estimate_min_f = 80,
+    .ff_estimate_max_f = 150,
 );
 
 PG_REGISTER_ARRAY(mixerRule_t, MIXER_RULE_COUNT, mixerRules, PG_GENERIC_MIXER_RULES, 0);